Remove initialData in nonce and session (#349)
It appears that in @tanstack/react-query v5 and up the query will not be fetched on mount if `initialData` is set. This means that the `ckSiweNonce` and `ckSiweSession` queries don't fire until their respective intervals have elapsed.

This commit removes the `initialData` for both of these queries in order to restore previous functionality.